fit(format): pocket lints

+++ test/utils/MockERC20PermitDAI.sol
@@ -8,11 +8,10 @@ contract MockERC20PermitDAI {
     // --- EIP712 niceties ---
     bytes32 public DOMAIN_SEPARATOR;
     // bytes32 public constant PERMIT_TYPEHASH = keccak256("Permit(address holder,address spender,uint256 nonce,uint256 expiry,bool allowed)");
-    bytes32 public constant PERMIT_TYPEHASH =
-        0xea2aa0a1be11a07ed86d755c93467f4f82362b452371d1ba94d1715123511acb;
+    bytes32 public constant PERMIT_TYPEHASH = 0xea2aa0a1be11a07ed86d755c93467f4f82362b452371d1ba94d1715123511acb;
     // --- Auth ---
 
-    mapping(address => uint) public wards;
+    mapping(address => uint256) public wards;
 
     function rely(address guy) external auth {
         wards[guy] = 1;
@@ -34,19 +33,19 @@ contract MockERC20PermitDAI {
     uint8 public constant decimals = 18;
     uint256 public totalSupply;
 
-    mapping(address => uint) public balanceOf;
-    mapping(address => mapping(address => uint)) public allowance;
-    mapping(address => uint) public nonces;
+    mapping(address => uint256) public balanceOf;
+    mapping(address => mapping(address => uint256)) public allowance;
+    mapping(address => uint256) public nonces;
 
-    event Approval(address indexed src, address indexed guy, uint wad);
-    event Transfer(address indexed src, address indexed dst, uint wad);
+    event Approval(address indexed src, address indexed guy, uint256 wad);
+    event Transfer(address indexed src, address indexed dst, uint256 wad);
 
     // --- Math ---
-    function add(uint x, uint y) internal pure returns (uint z) {
+    function add(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require((z = x + y) >= x);
     }
 
-    function sub(uint x, uint y) internal pure returns (uint z) {
+    function sub(uint256 x, uint256 y) internal pure returns (uint256 z) {
         require((z = x - y) <= x);
     }
 
@@ -54,9 +53,7 @@ contract MockERC20PermitDAI {
         wards[msg.sender] = 1;
         DOMAIN_SEPARATOR = keccak256(
             abi.encode(
-                keccak256(
-                    "EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)"
-                ),
+                keccak256("EIP712Domain(string name,string version,uint256 chainId,address verifyingContract)"),
                 keccak256(bytes(name)),
                 keccak256(bytes(version)),
                 getChainId(),
@@ -74,23 +71,14 @@ contract MockERC20PermitDAI {
     }
 
     // --- Token ---
-    function transfer(address dst, uint wad) external returns (bool) {
+    function transfer(address dst, uint256 wad) external returns (bool) {
         return transferFrom(msg.sender, dst, wad);
     }
 
-    function transferFrom(
-        address src,
-        address dst,
-        uint wad
-    ) public returns (bool) {
+    function transferFrom(address src, address dst, uint256 wad) public returns (bool) {
         require(balanceOf[src] >= wad, "Dai/insufficient-balance");
-        if (
-            src != msg.sender && allowance[src][msg.sender] != type(uint256).max
-        ) {
-            require(
-                allowance[src][msg.sender] >= wad,
-                "Dai/insufficient-allowance"
-            );
+        if (src != msg.sender && allowance[src][msg.sender] != type(uint256).max) {
+            require(allowance[src][msg.sender] >= wad, "Dai/insufficient-allowance");
             allowance[src][msg.sender] = sub(allowance[src][msg.sender], wad);
         }
         balanceOf[src] = sub(balanceOf[src], wad);
@@ -99,21 +87,16 @@ contract MockERC20PermitDAI {
         return true;
     }
 
-    function mint(address usr, uint wad) external auth {
+    function mint(address usr, uint256 wad) external auth {
         balanceOf[usr] = add(balanceOf[usr], wad);
         totalSupply = add(totalSupply, wad);
         emit Transfer(address(0), usr, wad);
     }
 
-    function burn(address usr, uint wad) external {
+    function burn(address usr, uint256 wad) external {
         require(balanceOf[usr] >= wad, "Dai/insufficient-balance");
-        if (
-            usr != msg.sender && allowance[usr][msg.sender] != type(uint256).max
-        ) {
-            require(
-                allowance[usr][msg.sender] >= wad,
-                "Dai/insufficient-allowance"
-            );
+        if (usr != msg.sender && allowance[usr][msg.sender] != type(uint256).max) {
+            require(allowance[usr][msg.sender] >= wad, "Dai/insufficient-allowance");
             allowance[usr][msg.sender] = sub(allowance[usr][msg.sender], wad);
         }
         balanceOf[usr] = sub(balanceOf[usr], wad);
@@ -121,22 +104,22 @@ contract MockERC20PermitDAI {
         emit Transfer(usr, address(0), wad);
     }
 
-    function approve(address usr, uint wad) external returns (bool) {
+    function approve(address usr, uint256 wad) external returns (bool) {
         allowance[msg.sender][usr] = wad;
         emit Approval(msg.sender, usr, wad);
         return true;
     }
 
     // --- Alias ---
-    function push(address usr, uint wad) external {
+    function push(address usr, uint256 wad) external {
         transferFrom(msg.sender, usr, wad);
     }
 
-    function pull(address usr, uint wad) external {
+    function pull(address usr, uint256 wad) external {
         transferFrom(usr, msg.sender, wad);
     }
 
-    function move(address src, address dst, uint wad) external {
+    function move(address src, address dst, uint256 wad) external {
         transferFrom(src, dst, wad);
     }
 
@@ -155,16 +138,7 @@ contract MockERC20PermitDAI {
             abi.encodePacked(
                 "\x19\x01",
                 DOMAIN_SEPARATOR,
-                keccak256(
-                    abi.encode(
-                        PERMIT_TYPEHASH,
-                        holder,
-                        spender,
-                        nonce,
-                        expiry,
-                        allowed
-                    )
-                )
+                keccak256(abi.encode(PERMIT_TYPEHASH, holder, spender, nonce, expiry, allowed))
             )
         );
 
@@ -172,7 +146,7 @@ contract MockERC20PermitDAI {
         require(holder == ecrecover(digest, v, r, s), "Dai/invalid-permit");
         require(expiry == 0 || block.timestamp <= expiry, "Dai/permit-expired");
         require(nonce == nonces[holder]++, "Dai/invalid-nonce");
-        uint wad = allowed ? type(uint256).max : 0;
+        uint256 wad = allowed ? type(uint256).max : 0;
         allowance[holder][spender] = wad;
         emit Approval(holder, spender, wad);
     }
